Tweak getStaticPaths of blog
